rename component from tabbedSubtitle to subtitle
