+++ app/components/views/TicketsPage/PurchaseTab/Tickets.js
@@ -3,7 +3,7 @@ import TicketAutoBuyer from "./TicketAutoBuyer";
 import { FormattedMessage as T } from "react-intl";
 import StakeInfo from "./StakeInfo";
 import { spv } from "connectors";
-import { ShowWarning, TabbedSubtitle } from "shared";
+import { ShowWarning, Subtitle } from "shared";
 import "style/StakePool.less";
 
 const Tickets = ({
@@ -12,7 +12,7 @@ const Tickets = ({
   ...props,
 }) => (
   <>
-    <TabbedSubtitle title={<T id="purchase.subtitle" m="Purchase Tickets"/>} />
+    <Subtitle title={<T id="purchase.subtitle" m="Purchase Tickets"/>} />
     <StakeInfo />
     {
       spvMode && blocksNumberToNextTicket == 2  ?
